# Simd aarch64 long seq opt