--- a/src/utils/packing/sse.rs
+++ b/src/utils/packing/sse.rs
@@ -130,3 +130,26 @@ pub fn as_2bit(seq: &[u8]) -> Result<u64, NucleotideError> {
 
     Ok(packed)
 }
+
+pub fn encode_internal(sequence: &[u8], ebuf: &mut Vec<u64>) -> Result<(), NucleotideError> {
+    // Clear the buffer
+    ebuf.clear();
+
+    // Calculate the number of chunks
+    let n_chunks = sequence.len().div_ceil(32);
+
+    let mut l_bounds = 0;
+    for _ in 0..n_chunks - 1 {
+        let r_bounds = l_bounds + 32;
+        let chunk = &sequence[l_bounds..r_bounds];
+
+        let bits = as_2bit(chunk)?;
+        ebuf.push(bits);
+        l_bounds = r_bounds;
+    }
+
+    let bits = as_2bit(&sequence[l_bounds..])?;
+    ebuf.push(bits);
+
+    Ok(())
+}