This PR is a follow up on: https://github.com/ArcInstitute/binseq/pull/36

This PR adds an optimized implementation for encode and decode that show a meaningful speedup on longer sequences

### Benching (updated)

Bench on long context before changes

```bash
cargo bench
```

output
```
     Running benches/packing_benchmark.rs (target/release/deps/packing_benchmark-f408d428e8068bfb)
packing/pack_simd/1     time:   [15.703 ns 15.718 ns 15.738 ns]
                        change: [-1.5072% -1.1275% -0.7000%] (p = 0.00 < 0.05)
                        Change within noise threshold.
Found 16 outliers among 100 measurements (16.00%)
  3 (3.00%) high mild
  13 (13.00%) high severe
packing/pack_simd/17    time:   [24.656 ns 24.736 ns 24.816 ns]
                        change: [+0.1503% +0.6319% +1.1098%] (p = 0.01 < 0.05)
                        Change within noise threshold.
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) low mild
packing/pack_simd/64    time:   [56.342 ns 56.398 ns 56.455 ns]
                        change: [-1.7254% -1.0753% -0.4449%] (p = 0.00 < 0.05)
                        Change within noise threshold.
Found 16 outliers among 100 measurements (16.00%)
  11 (11.00%) high mild
  5 (5.00%) high severe
packing/pack_simd/128   time:   [91.771 ns 91.897 ns 92.035 ns]
                        change: [-0.2002% +0.2245% +0.6713%] (p = 0.28 > 0.05)
                        No change in performance detected.
Found 11 outliers among 100 measurements (11.00%)
  5 (5.00%) high mild
  6 (6.00%) high severe
packing/pack_simd/256   time:   [186.01 ns 186.23 ns 186.51 ns]
                        change: [-2.5612% -1.5883% -0.7292%] (p = 0.00 < 0.05)
                        Change within noise threshold.
Found 14 outliers among 100 measurements (14.00%)
  2 (2.00%) high mild
  12 (12.00%) high severe
packing/pack_simd/512   time:   [351.39 ns 351.60 ns 351.85 ns]
                        change: [+0.0867% +0.2888% +0.4804%] (p = 0.00 < 0.05)
                        Change within noise threshold.
Found 10 outliers among 100 measurements (10.00%)
  3 (3.00%) high mild
  7 (7.00%) high severe
packing/pack_simd/1024  time:   [655.76 ns 656.15 ns 656.57 ns]
                        change: [-0.5243% -0.3303% -0.1318%] (p = 0.00 < 0.05)
                        Change within noise threshold.
Found 12 outliers among 100 measurements (12.00%)
  1 (1.00%) low severe
  3 (3.00%) high mild
  8 (8.00%) high severe
packing/pack_simd/32000 time:   [18.049 µs 18.069 µs 18.092 µs]
                        change: [+0.0917% +0.3060% +0.5119%] (p = 0.00 < 0.05)
                        Change within noise threshold.
Found 6 outliers among 100 measurements (6.00%)
  6 (6.00%) high mild
packing/pack_simd/64000 time:   [35.746 µs 35.834 µs 35.935 µs]
                        change: [-0.3255% +0.0171% +0.3191%] (p = 0.92 > 0.05)
                        No change in performance detected.
Found 27 outliers among 100 measurements (27.00%)
  10 (10.00%) low mild
  3 (3.00%) high mild
  14 (14.00%) high severe
packing/pack_simd/128000
                        time:   [71.363 µs 71.600 µs 71.956 µs]
                        change: [-0.0679% +0.2348% +0.5526%] (p = 0.15 > 0.05)
                        No change in performance detected.
Found 24 outliers among 100 measurements (24.00%)
  5 (5.00%) low severe
  6 (6.00%) low mild
  3 (3.00%) high mild
  10 (10.00%) high severe
packing/pack_simd/256000
                        time:   [142.73 µs 142.82 µs 142.93 µs]
                        change: [-0.1243% +0.0550% +0.2286%] (p = 0.55 > 0.05)
                        No change in performance detected.
Found 13 outliers among 100 measurements (13.00%)
  2 (2.00%) high mild
  11 (11.00%) high severe
packing/pack_simd/512000
                        time:   [290.82 µs 291.17 µs 291.50 µs]
                        change: [-0.6153% -0.2769% +0.0337%] (p = 0.10 > 0.05)
                        No change in performance detected.
Found 8 outliers among 100 measurements (8.00%)
  6 (6.00%) high mild
  2 (2.00%) high severe

     Running benches/sequence_benchmark.rs (target/release/deps/sequence_benchmark-0883711688bc8d3a)
```

and after 

```bash
cargo bench
```

after

```
     Running benches/packing_benchmark.rs (target/release/deps/packing_benchmark-f408d428e8068bfb)
packing/pack_simd/1     time:   [15.426 ns 15.473 ns 15.529 ns]
                        change: [-2.2511% -1.7766% -1.3426%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 12 outliers among 100 measurements (12.00%)
  7 (7.00%) high mild
  5 (5.00%) high severe
packing/pack_simd/17    time:   [24.441 ns 24.509 ns 24.601 ns]
                        change: [-0.4958% +0.1656% +0.8013%] (p = 0.61 > 0.05)
                        No change in performance detected.
packing/pack_simd/64    time:   [36.401 ns 36.687 ns 36.917 ns]
                        change: [-35.356% -35.030% -34.740%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 11 outliers among 100 measurements (11.00%)
  4 (4.00%) low severe
  2 (2.00%) low mild
  5 (5.00%) high mild
packing/pack_simd/128   time:   [42.206 ns 42.914 ns 43.623 ns]
                        change: [-53.628% -53.150% -52.698%] (p = 0.00 < 0.05)
                        Performance has improved.
packing/pack_simd/256   time:   [60.647 ns 61.015 ns 61.378 ns]
                        change: [-67.200% -67.076% -66.959%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 23 outliers among 100 measurements (23.00%)
  5 (5.00%) low severe
  4 (4.00%) low mild
  12 (12.00%) high mild
  2 (2.00%) high severe
packing/pack_simd/512   time:   [71.757 ns 71.824 ns 71.902 ns]
                        change: [-79.578% -79.539% -79.502%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 33 outliers among 100 measurements (33.00%)
  23 (23.00%) low severe
  2 (2.00%) high mild
  8 (8.00%) high severe
packing/pack_simd/1024  time:   [128.82 ns 128.89 ns 129.00 ns]
                        change: [-80.398% -80.359% -80.320%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 15 outliers among 100 measurements (15.00%)
  1 (1.00%) low severe
  3 (3.00%) high mild
  11 (11.00%) high severe
packing/pack_simd/32000 time:   [4.1034 µs 4.1156 µs 4.1281 µs]
                        change: [-77.429% -77.351% -77.279%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 2 outliers among 100 measurements (2.00%)
  2 (2.00%) low mild
packing/pack_simd/64000 time:   [7.5325 µs 7.5491 µs 7.5658 µs]
                        change: [-78.931% -78.859% -78.784%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 5 outliers among 100 measurements (5.00%)
  5 (5.00%) high mild
packing/pack_simd/128000
                        time:   [14.700 µs 14.708 µs 14.717 µs]
                        change: [-79.468% -79.402% -79.345%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 15 outliers among 100 measurements (15.00%)
  4 (4.00%) low mild
  3 (3.00%) high mild
  8 (8.00%) high severe
packing/pack_simd/256000
                        time:   [28.782 µs 28.794 µs 28.809 µs]
                        change: [-79.889% -79.851% -79.815%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 10 outliers among 100 measurements (10.00%)
  1 (1.00%) high mild
  9 (9.00%) high severe
packing/pack_simd/512000
                        time:   [57.578 µs 57.703 µs 57.860 µs]
                        change: [-80.215% -80.159% -80.093%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 20 outliers among 100 measurements (20.00%)
  1 (1.00%) low mild
  7 (7.00%) high mild
  12 (12.00%) high severe

     Running benches/sequence_benchmark.rs (target/release/deps/sequence_benchmark-0883711688bc8d3a)
```


summary table of the performance improvements shown in the benchmark results

| Sequence Length | Before (ns/µs) | After (ns/µs) | Improvement (%) | Speedup (×) |
| --------------- | -------------- | ------------- | --------------- | ----------- |
| 1               | 15.718 ns      | 15.473 ns     | 1.56%           | 1.02×       |
| 17              | 24.736 ns      | 24.509 ns     | 0.92%           | 1.01×       |
| 64              | 56.398 ns      | 36.687 ns     | 34.95%          | 1.54×       |
| 128             | 91.897 ns      | 42.914 ns     | 53.30%          | 2.14×       |
| 256             | 186.23 ns      | 61.015 ns     | 67.24%          | 3.05×       |
| 512             | 351.60 ns      | 71.824 ns     | 79.57%          | 4.90×       |
| 1,024           | 656.15 ns      | 128.89 ns     | 80.36%          | 5.09×       |
| 32,000          | 18.069 µs      | 4.1156 µs     | 77.23%          | 4.39×       |
| 64,000          | 35.834 µs      | 7.5491 µs     | 78.93%          | 4.75×       |
| 128,000         | 71.600 µs      | 14.708 µs     | 79.46%          | 4.87×       |
| 256,000         | 142.82 µs      | 28.794 µs     | 79.84%          | 4.96×       |
| 512,000         | 291.17 µs      | 57.703 µs     | 80.18%          | 5.05×       |


These changes result in approximately 79-80% performance improvement on the longer sequences

Please let me know if I should make any changes! 🙏