@noamteyssier 2025-04-21
awesome thanks for the p.r! Do you think the same algorithm would work well for shorter sequences as well or is this limited to just long-reads?