@drbh 2025-04-22
Hey @noamteyssier! After a bit more testing it seems that for any sequence with length <32 it is significantly faster to use the current `as_2bit` function (~10-15ns faster in the case of seq len 1)

I've moved the branching logic into a new `encode_internal` function that each packing implementation now exports to keep the special case contained to the `aarch64.rs`.  

`cargo test` are passing and the benches are stable, however please let me know if I should restructure the changes in any way! 

Thanks for the quick responses 🙏